// src/app/shared/model/core.model.ts
export interface ApiResponseModel<T = any> {
	MRData: T;
}

export interface Location {
	lat: string;
	long: string;
	locality: string;
	country: string;
}

export interface Circuit {
	circuitId: string;
	url: string;
	circuitName: string;
	Location: Location;
}

export interface Result {
	number: string;
	position: string;
	positionText: string;
	points: string;
	Driver: Driver;
	Constructor: Constructor;
	grid: string;
	laps: string;
	status: string;
	Time: Time;
	FastestLap: FastestLap;
}

export interface Race {
	Results?: Result[];
	season: string;
	round: string;
	url: string;
	raceName: string;
	Circuit: Circuit;
	date: string;
	time: string;
}

export interface Driver {
	driverId: string;
	permanentNumber: string;
	code: string;
	url: string;
	givenName: string;
	familyName: string;
	dateOfBirth: string;
	nationality: string;
}

export interface Constructor {
	constructorId: string;
	url: string;
	name: string;
	nationality: string;
}

export interface FastestLap {
	rank: string;
	lap: string;
	Time: Time;
	AverageSpeed: AverageSpeed;
}

export interface Time {
	time: string;
	milis?: string;
}

export interface AverageSpeed {
	units: string;
	speed: string;
}